feat: move mappings to class
